pyproject.toml
[tool.poetry]
name = "business-backend"
version = "0.1.0"
description = "Business Backend with FastAPI, GraphQL, and LangChain"
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
strawberry-graphql = {extras = ["fastapi"], version = "^0.219.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
asyncpg = "^0.29.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
aioinject = "^0.32.0"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
langchain = "^0.1.0"
langchain-core = "^0.1.10"
langchain-openai = "^0.0.5"
pandas = "^2.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.4"
black = "^24.1.1"
isort = "^5.13.2"
mypy = "^1.8.0"
ipykernel = "^6.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
